+++ Source/ACE.Server/WorldObjects/Creature_Properties.cs
@@ -89,6 +89,12 @@ namespace ACE.Server.WorldObjects
             set { if (!value.HasValue) RemoveProperty(PropertyFloat.ResistNether); else SetProperty(PropertyFloat.ResistNether, value.Value); }
         }
 
+        public bool NonProjectileMagicImmune
+        {
+            get => GetProperty(PropertyBool.NonProjectileMagicImmune) ?? false;
+            set { if (!value) RemoveProperty(PropertyBool.NonProjectileMagicImmune); else SetProperty(PropertyBool.NonProjectileMagicImmune, value); }
+        }
+
         public float GetResistanceMod(DamageType damageType, WorldObject weapon, float weaponResistanceMod = 1.0f)
         {
             var ignoreMagicResist = weapon != null ? weapon.IgnoreMagicResist : false;